+++ programs/nlink/include/GsEdge.h
