+++ programs/nlink/include/GsNode.h
